half-baked update to CANcli
